requirements.txt
gradio>=4.44.0
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.10.0.84
numpy>=1.26.0
Pillow>=10.0.0
scikit-image>=0.24.0
open3d>=0.18.0
tqdm>=4.66.0
huggingface_hub>=0.24.0
